pyproject.toml
[project]
name = "autorefine"
version = "0.1.0"
description = "Generalized auto-refinement loop for text artifacts using LLM-as-judge evaluation"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
